requirements.txt
Pillow
numpy==1.15.0
